+++ feign-core/src/main/java/feign/FeignException.java
@@ -32,8 +32,8 @@ public class FeignException extends RuntimeException {
   public static FeignException errorStatus(String methodKey, Response response) {
     String message = format("status %s reading %s", response.status(), methodKey);
     try {
-      Object body = toString.decode(response, String.class);
-      if (body != null) {
+      if (response.body() != null) {
+        String body = toString.decode(response.body().asReader(), String.class);
         response =
             Response.create(
                 response.status(), response.reason(), response.headers(), body.toString());