+++ feign-core/src/test/java/feign/LoggerTest.java
@@ -15,20 +15,18 @@
  */
 package feign;
 
+import static dagger.Provides.Type.SET;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
-import com.google.common.collect.ImmutableMap;
 import com.google.mockwebserver.MockResponse;
 import com.google.mockwebserver.MockWebServer;
 import dagger.Provides;
-import feign.codec.Decoder;
-import feign.codec.StringDecoder;
+import feign.codec.Encoder;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import org.testng.annotations.BeforeMethod;
@@ -112,10 +110,14 @@ public class LoggerTest {
 
     @dagger.Module(overrides = true, library = true)
     class Module {
-      @Provides
-      @Singleton
-      Map<String, Decoder> decoders() {
-        return ImmutableMap.<String, Decoder>of("SendsStuff", new StringDecoder());
+      @Provides(type = SET)
+      Encoder defaultEncoder() {
+        return new Encoder.Text<Object>() {
+          @Override
+          public String encode(Object object) {
+            return object.toString();
+          }
+        };
       }
 
       @Provides