+++ feign-core/src/test/java/feign/codec/DefaultErrorDecoderTest.java
@@ -26,6 +26,8 @@ import java.util.Collection;
 import org.testng.annotations.Test;
 
 public class DefaultErrorDecoderTest {
+  ErrorDecoder errorDecoder = new ErrorDecoder.Default();
+
   @Test(
       expectedExceptions = FeignException.class,
       expectedExceptionsMessageRegExp = "status 500 reading Service#foo\\(\\)")
@@ -34,7 +36,7 @@ public class DefaultErrorDecoderTest {
         Response.create(
             500, "Internal server error", ImmutableMap.<String, Collection<String>>of(), null);
 
-    throw ErrorDecoder.DEFAULT.decode("Service#foo()", response);
+    throw errorDecoder.decode("Service#foo()", response);
   }
 
   @Test(
@@ -49,7 +51,7 @@ public class DefaultErrorDecoderTest {
             ImmutableMap.<String, Collection<String>>of(),
             "hello world");
 
-    throw ErrorDecoder.DEFAULT.decode("Service#foo()", response);
+    throw errorDecoder.decode("Service#foo()", response);
   }
 
   @Test(
@@ -63,6 +65,6 @@ public class DefaultErrorDecoderTest {
             ImmutableMultimap.of(RETRY_AFTER, "Sat, 1 Jan 2000 00:00:00 GMT").asMap(),
             null);
 
-    throw ErrorDecoder.DEFAULT.decode("Service#foo()", response);
+    throw errorDecoder.decode("Service#foo()", response);
   }
 }