+++ feign-jaxrs/src/test/java/feign/jaxrs/examples/GitHubExample.java
@@ -15,18 +15,19 @@
  */
 package feign.jaxrs.examples;
 
-import com.google.common.collect.ImmutableMap;
+import static dagger.Provides.Type.SET;
+
 import com.google.gson.Gson;
+import com.google.gson.JsonIOException;
 import dagger.Module;
 import dagger.Provides;
 import feign.Feign;
 import feign.codec.Decoder;
 import feign.jaxrs.JAXRSModule;
+import java.io.IOException;
 import java.io.Reader;
 import java.lang.reflect.Type;
 import java.util.List;
-import java.util.Map;
-import javax.inject.Singleton;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -59,20 +60,23 @@ public class GitHubExample {
   /** JAXRSModule tells us to process @GET etc annotations */
   @Module(overrides = true, library = true, includes = JAXRSModule.class)
   static class GitHubModule {
-    @Provides
-    @Singleton
-    Map<String, Decoder> decoders() {
-      return ImmutableMap.of("GitHub", jsonDecoder);
-    }
+    @Provides(type = SET)
+    Decoder decoder() {
+      return new Decoder.TextStream<Object>() {
+        Gson gson = new Gson();
 
-    final Decoder jsonDecoder =
-        new Decoder() {
-          Gson gson = new Gson();
-
-          @Override
-          public Object decode(Reader reader, Type type) {
+        @Override
+        public Object decode(Reader reader, Type type) throws IOException {
+          try {
             return gson.fromJson(reader, type);
+          } catch (JsonIOException e) {
+            if (e.getCause() != null && e.getCause() instanceof IOException) {
+              throw IOException.class.cast(e.getCause());
+            }
+            throw e;
           }
-        };
+        }
+      };
+    }
   }
 }