+++ feign-jaxrs/src/test/java/feign/jaxrs/examples/IAMExample.java
@@ -15,7 +15,8 @@
  */
 package feign.jaxrs.examples;
 
-import com.google.common.collect.ImmutableMap;
+import static dagger.Provides.Type.SET;
+
 import dagger.Module;
 import dagger.Provides;
 import feign.Feign;
@@ -26,8 +27,6 @@ import feign.codec.Decoder;
 import feign.codec.Decoders;
 import feign.examples.AWSSignatureVersion4;
 import feign.jaxrs.JAXRSModule;
-import java.util.Map;
-import javax.inject.Singleton;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 
@@ -75,10 +74,9 @@ public class IAMExample {
 
   @Module(overrides = true, library = true, includes = JAXRSModule.class)
   static class IAMModule {
-    @Provides
-    @Singleton
-    Map<String, Decoder> decoders() {
-      return ImmutableMap.of("IAM#arn()", Decoders.firstGroup("<Arn>([\\S&&[^<]]+)</Arn>"));
+    @Provides(type = SET)
+    Decoder decoder() {
+      return Decoders.firstGroup("<Arn>([\\S&&[^<]]+)</Arn>");
     }
   }
 }