+++ CHANGES.md
@@ -1,8 +1,14 @@
 ### Version 3.0
 * Wire is now Logger, with configurable Logger.Level.
-* decoupled ErrorDecoder from fallback handling
-* Decoders can throw checked exceptions, but needn't declare Throwable
-* Decoders no longer read methodKey
+* changed codec to be similar to [WebSocket JSR 356](http://docs.oracle.com/javaee/7/api/javax/websocket/package-summary.html)
+  * Decoder is now `Decoder.TextStream<T>`
+  * BodyEncoder is now `Encoder.Text<T>`
+  * FormEncoder is now `Encoder.Text<Map<String, ?>>`
+* Encoder and Decoders are specified via `Provides.Type.SET` binding.
+* Default Encoder and Form Encoder is `Encoder.Text<Object>`
+* Default Decoder is `Decoder.TextStream<Object>`
+* ErrorDecoder now returns Exception, not fallback.
+* There can only be one `ErrorDecoder` and `Request.Options` binding now. 
 
 ### Version 2.0.0
 * removes guava and jax-rs dependencies